requirements.txt
Pillow==11.3.0
torch==1.12.1+cu113
torchvision==0.13.1+cu113
